+++ view/next-project/src/components/common/RecordEditModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RecordEditModal';
